feat: add game win screen and victory condition Ball: - Add noexcept to posY/posX getters - Adjust ball speed (75ms -> 105ms)
Blocks:
- Add isAllBlocksDestroyed() and generateBlocksPattern() methods
- Add generalBlocksAmount_ to track total blocks
- Add noexcept to destroyedCount()

Game:
- Add GameWinScreen member
- Render win screen on all blocks destroyed
- Update exit condition for both game over and win
- Stop game on ball lost OR victory

ConsoleViewport:
- Adjust viewport width (2x -> 3x margin)

CMakeLists.txt:
- Add GameWinScreen.cpp

+++ include/GameWindow.hpp
@@ -31,10 +31,9 @@ private:
     static constexpr int PANEL_X_DIVISOR = 2;
 
     static constexpr int FIELD_BORDER_OFFSET = 2;
-    static constexpr int FIELD_WIDTH_SCALE = 3;
+    static constexpr double FIELD_WIDTH_SCALE = 3;
     static constexpr int FIELD_Y_OFFSET = 1;
     static constexpr int FIELD_X_DIVISOR = 6;
-
 
     int height_;
     int width_;
